Added pdf-service
Service runs at port 9030 and listens to all POST requests
to /generate?stylesheet=xyz, where xyz is the basename of an
installed .XSL file. The request body must contain a JSON object.
The object is then converted to XML document that uses the given .XSL file
as stylesheet. This document is converted to PDF and the result
is returned.