allu-ui-service: APIs to generate and fetch decision
Added the following new APIs:

* PUT /applications/{applicationId}/decision -- generate decision
* GET /applications/{applicationId/decision  -- fetch decision

In this commit tha APIs are half-dummy, i.e. generate does
nothing and fetch always generates a new decision. The idea
behind this implementation is that the web fron can already
use these APIS in the proper generate-fetch sequence even
though the actual storing of generated PDF is not implemented yet.

Also fixed a TODO in ApplicationService.