backend: add type property to Event classes
For Jackson serialization to work for class hierarchies,
the serializer had been configured to add a virtual
"type" property to the JSON string while serializing
Event classes. Now the same property is useful also
in the Java code that handles various events and user
permissions, so it was implemented as a proper
read-only property.