model-service: added check for matches in geometry insert
LocationGeometryInserter now generates SQL code that checks
that the given name doesn't match multiple different names in
the database and that it still matches one name. If the checks
fail, a SQL exception is raised. This makes it possible to
catch errors in location names already during build time as
"mvn generate-sources" will fail.

Also added missing locations for Asema-aukio A & B and fixed
a double name for bridge 18a. These problems were detected as
a result of the new SQL code.