backend: fix bug with same customer in multiple roles
If the same customer was in many roles in the same application,
and had the same contat in different roles, the saving the
application failed, since the CustomerWithContacts data in the
saved application would contain the customer with multiple
copies of its contact. This commit fixes the errorous DAO
code.