Backend: fix cable report mapping when same contact twice as orderer
A contact can be multiple times in a cable report, as a single customer
can appear in multiple roles and give contacts for each role. Therefore
it's possible that the orderer contact may appear as contact for many
customer roles and that should not be handled as error: it is still
the one and same customer.

Fixed the checking code.