backend: added id field to CustomerWithContactsJson
The 'id' field is read-only and always the same as the customer's id.
It is only needed in application history generation where two lists
of CustomerWithContactJson objects are compared. If the objects don't
have 'id' field they are compared based on their index in the parent
list, which gives more or less random results.