Implement actions and effects for decision termination
The effects of saving the termination info and moving the
termination to decision phase will be chained together if the
termination is to be moved directly to decision phase.

Some currently unused actions and states are implemented, but
these are expected to be of use when implementing loading existing
termination info for the modal and the tab for decision document.

The service and backend implementation for moving the decision is
still missing so it's simulated by a delayed emit.