Match incoming customer data primarily by sap customer number
Add sap customer number to queries to use it as a search condition.
Add boost parameter to prioritize query parameters.

The query builder is changed from .filter() to .must() to enable
scoring as filter context removes scoring info. The actual results are
the same as far as I understand.

Query parameters that have a boost parameter assigned, will be handled
as a constant score query, as it is the simplest way to prioritize
multiple search parameters.

The method searchCustomer in CustomerAcceptanceComponent was refactored
for readability. The functionality of registry key parameter should
be unchanged.