Require orderer for cable report
Orderer field is required and it is not set as default anymore since this should be users choice.
Orderer missing is shown as error notification since it does not belong to any single field or element in
application form.

In case of orderer missing and other errors show generic error message with orderer error.