Pdf recognition when dragging attachment to drop area
Dragging a pdf file to drop down area in attachments list did not recognize whether attachment
was pdf or not. Only pdf's are eligible for decision attachments so this was a problem.

Bug was fixed by setting the mime type of the attachment after file drop and check
file type when initializing the actual attachment component which shows the attachment details.