Separate cards for invoicing details and deposit
Invoicing details and deposit were in the same card previously which was messy
and deposit is not shown even for all application types.

Restructured the layout to contain two separate cards, one for invoicing details, one
for deposit aligned in column.

When not-invoiceable is selected cards are aligned as a row.