Add logging in case of exception in initial search data sync
Now you don't have to keep wondering why the sync isn't going through
when you've forgotten to start the api for external services.

ScheduledExecutorService suppresses all exceptions to handle them
in a different way. Simplest way to log this exception is to catch
it in the runner instead of using the intended mechanism via .get()