Fix for workqueue sort changing when also having stored filter
There is currently no way to set specific sort programmatically in MatSort.
Because of this we need to clear active sort and set a new sort to prevent
unwanted behaviour.

There is still a bug where sort arrow does not appear when setting sort programmatically
when there is no sort arrow visible. This is a known bug in material and propably
will be fixed in future.