External Api for reporting application changes without information request
There are cases where customer wants to report application changes even to ongoing applications.
This new api provides a way to report application changes to applicion which is in decision or operation condition status.
Implementation creates a simple information request and adds reported changes as response for it.

This allows similar handling as normal information request - reponse flow.