Implement endpoint for updating locations in supervision api
Implemented corresponding endpoint in model service. LocationDao::update
was previously only used via updating all locations of an application
simultaneously so it was made public to enable single location update.

As stated in the comments, ModelFieldUpdater required a workaround to
work propely with Geometry classes. The SimpleModule is required as
the existing usage of the (de)serializer proxy classes use property
annotations in Location class. These are not used when converting
a Geometry value outside of the Location class.