Removed unwanted map search from application basic info
When application basic info was opened ui made a application search based on visible area shown on map.
This is not wanted since only selected application's geometry is shown on the map and that is already known.
Fix for this was to change map components role to OTHER since that role does not do map searches.

Also fixed a bug where map searches were triggered before effects which handle them were initialized. This
caused search actions to be dispatched but responses never came. This caused showing loading icon until
actual loading was done later.