Handle application replacement in information acceptance
When application is replaced information acceptance should be repopulated with
replacing applications information. This is done via setting replacing application to oldInfo.
Some operations in child components are not ran with new info but it shouldn't be a problem
since replacement does not update fields which are handled in acceptance.

Location updates are handled a bit differently. On application replace locations
are set from new application. After this, all updates to locations are handled via
location key. This is done so locations' id's don't change after they are set
from replacing application.