Use invoice recipient from invoice when available
Invoice recipient is a snapshot of invoice customer taken when invoice is created. Invoice
recipient info should be shown from this snapshot when available since it is the state of
customer when invoice was made.

Invoice recipient is not available before invoice creation (decision) so application's
linked invoice recipient (customer) is used as a fallback.