Handle exceptions when sending reminder emails
Catch exceptions when sending reminder emails and log them to mail sender log. This allows
other reminders to be sent and marked as sent without failing. When sending failed with
old implementation none was marked as sent. This caused resend for all of those again and again.

New implementation handles error in message sending. This can cause a situation where
all reminders are marked as sent although some failed. This is acceptable though since failures
are logged in database so they can be identified.