Make use of Docker container aliases
Using the Docker network gateway as a common
IP address to route all traffic using only IPs
did not succeed. Thus, another solution needs to
be implemented.
As Java8 URI could not handle underscores in
hostnames, the names should be okay to use, as
long as underscores are avoided.
Keep Docker network configs to ensure that
firewall settings are set correctly.