fixed swagger documentation behind proxy
Needed to add X-Forwarded-Prefix on requests that were forwarded from frontend to backend, so that swagger knows correct path to generate for the documentation.
To achieve this a forward header state needs to be added on spring. Caused some problems with spring security. Even though servletPath is correct when calling controllers, spring security will read call with reverse proxy forwarding path. Because of that, paths need to be added to access-path that also have reverse proxy prefix.
Removed unnecessary resources that are not used.