ALLU-116: Don't archive cable reports with active excavation announcements
Check whether an otherwise archivable cable report is associated with
any still active excavation announcements, and prevent archiving if
it does. Implement the archivability logic and a mocked test, but
the actual fetching of active excavation announcements is still
unimplemented in this commit.

I'm not really happy with how the fetching and releasing the active
excavation announcements works here, but since the existing process
only fetches applications one by one as it iterates through them,
we can't know beforehand if we need the excavation announcements,
and we don't know when we can release them other than at the end
of the whole archival process. At least fetching them is now
maximally lazy, as they are only fetched if all the other
conditions for archiving a cable report are fulfilled first.

(It would also be nicer if we could just get a boolean back from
the database, but since QueryDSL can't handle the JSON objects
inside the database, we have to do it this way by naively fetching
all active excavation announcements and checking them manually.
Since this is called once a day in the background by the
scheduler-service, it doesn't really matter if it is somewhat slower
than it might be in the best possible world.)