ALLU-85: Implement the cable report anonymization logic
Implement a stubbed version of the cable report anonymization logic
and a mocked test for it. This relies on the yet unimplemented
fetching of the potentially anonymizable cable reports to fetch
only those applications for which those anonymization conditions
which can be tested for directly in the database query are fulfilled.