ALLU-125: Fix problem in anonymizing replaced applications
The removeAdditionalInfos() used as part of the anonymization
process called ApplicationDao.update() which wants to refresh the
application data from the DB after the update. However it uses
the findByIds() function which does not return applications whose
status is REPLACED, so this caused update() to throw an Index out
of bounds exception when anonymizing replaced applications. Fix
by splitting ApplicationDao.update() into two parts and using
the part which doesn't fetch the new data in removeAdditionalInfos()